+++ ios/disk_space_plus/Sources/disk_space_plus/DiskSpacePlusPlugin.m
@@ -1,5 +1,9 @@
 #import "DiskSpacePlusPlugin.h"
+#if __has_include(<disk_space_plus/disk_space_plus-Swift.h>)
 #import <disk_space_plus/disk_space_plus-Swift.h>
+#else
+#import "disk_space_plus-Swift.h"
+#endif
 
 @implementation DiskSpacePlusPlugin
 + (void)registerWithRegistrar:(NSObject<FlutterPluginRegistrar>*)registrar {